requirements.txt
requests>=2.32.0
python-dotenv>=1.0.1
pydantic>=2.7.0
lxml>=5.3.0
tenacity>=9.0.0
fastapi>=0.115.0
uvicorn>=0.30.0

